# feat(db): use write buffer to speed up state migrations