--- a/src/tool/subcommands/shed_cmd/migration.rs
+++ b/src/tool/subcommands/shed_cmd/migration.rs
@@ -7,8 +7,10 @@ use std::time::Instant;
 
 use cid::Cid;
 use clap::Args;
+use fvm_ipld_blockstore::Blockstore;
 use itertools::Itertools;
 
+use crate::state_migration::BlockstoreWithWriteBuffer;
 use crate::utils::db::CborStoreExt;
 use crate::{
     blocks::CachingBlockHeader,
@@ -36,6 +38,9 @@ pub struct MigrateStateCommand {
     /// Filecoin network chain
     #[arg(long, required = true)]
     chain: NetworkChain,
+    /// Size of database write buffer, use 0 to disable write buffer
+    #[arg(long, default_value_t = 10000)]
+    db_write_buffer: usize,
 }
 
 impl MigrateStateCommand {
@@ -45,6 +50,7 @@ impl MigrateStateCommand {
             block_to_look_back,
             db,
             chain,
+            db_write_buffer,
         } = self;
         let db = {
             let db = if let Some(db) = db {
@@ -53,7 +59,15 @@ impl MigrateStateCommand {
                 let (_, config) = read_config(None, Some(chain.clone()))?;
                 db_root(&chain_path(&config))?
             };
-            load_db(&db)?
+            let db = load_db(&db)?;
+            Arc::new(if db_write_buffer > 0 {
+                Either::Left(BlockstoreWithWriteBuffer::new_with_capacity(
+                    db,
+                    db_write_buffer,
+                ))
+            } else {
+                Either::Right(db)
+            })
         };
         let block: CachingBlockHeader = db.get_cbor_required(&block_to_look_back)?;
         let chain_config = Arc::new(ChainConfig::from_chain(&chain));
@@ -90,4 +104,73 @@ pub(super) fn load_db(db_root: &Path) -> anyhow::Result<Arc<ManyCar<ParityDb>>> 
     let forest_car_db_dir = db_root.join(CAR_DB_DIR_NAME);
     load_all_forest_cars(&db, &forest_car_db_dir)?;
     Ok(Arc::new(db))
+}
+
+enum Either<A: Blockstore, B: Blockstore> {
+    Left(A),
+    Right(B),
+}
+
+impl<A: Blockstore, B: Blockstore> Blockstore for Either<A, B> {
+    fn has(&self, k: &Cid) -> anyhow::Result<bool> {
+        match self {
+            Self::Left(v) => v.has(k),
+            Self::Right(v) => v.has(k),
+        }
+    }
+
+    #[allow(clippy::disallowed_types)]
+    fn put<D>(
+        &self,
+        mh_code: multihash_codetable::Code,
+        block: &fvm_ipld_blockstore::Block<D>,
+    ) -> anyhow::Result<Cid>
+    where
+        Self: Sized,
+        D: AsRef<[u8]>,
+    {
+        match self {
+            Self::Left(v) => v.put(mh_code, block),
+            Self::Right(v) => v.put(mh_code, block),
+        }
+    }
+
+    #[allow(clippy::disallowed_types)]
+    fn put_many<D, I>(&self, blocks: I) -> anyhow::Result<()>
+    where
+        Self: Sized,
+        D: AsRef<[u8]>,
+        I: IntoIterator<Item = (multihash_codetable::Code, fvm_ipld_blockstore::Block<D>)>,
+    {
+        match self {
+            Self::Left(v) => v.put_many(blocks),
+            Self::Right(v) => v.put_many(blocks),
+        }
+    }
+
+    fn put_many_keyed<D, I>(&self, blocks: I) -> anyhow::Result<()>
+    where
+        Self: Sized,
+        D: AsRef<[u8]>,
+        I: IntoIterator<Item = (Cid, D)>,
+    {
+        match self {
+            Self::Left(v) => v.put_many_keyed(blocks),
+            Self::Right(v) => v.put_many_keyed(blocks),
+        }
+    }
+
+    fn get(&self, k: &Cid) -> anyhow::Result<Option<Vec<u8>>> {
+        match self {
+            Self::Left(v) => v.get(k),
+            Self::Right(v) => v.get(k),
+        }
+    }
+
+    fn put_keyed(&self, k: &Cid, block: &[u8]) -> anyhow::Result<()> {
+        match self {
+            Self::Left(v) => v.put_keyed(k, block),
+            Self::Right(v) => v.put_keyed(k, block),
+        }
+    }
 }