## Summary of changes

<!-- Please write a comprehensive summary of your changes and what was the motivation behind them -->

As discussed on [slack](https://filecoinproject.slack.com/archives/C029LPZ5N73/p1744674781996969?thread_ts=1744670553.976459&cid=C029LPZ5N73), this PR implements write buffer to speed up state migrations

On my laptop, nv25 mainnet upgrade takes 5m with 10k buffer, and 22m without buffer.
```
➜  forest git:(main) ✗ forest-tool shed migrate-state --chain mainnet 25 bafy2bzacebb44z6iy77lzpkdxosxk5c2rxtyststp6ukvbknfokswi4czrr3i
Migrating... state_root: bafy2bzaced5lda3j6wpzmgglta47xcxdmnxhanxyiugdme4bl3fdieiaot6go, epoch: 4878799
2025-04-21T11:47:32.422007Z  INFO forest::state_migration::common::state_migration: Processed 100000 actors
...
2025-04-21T11:51:27.950242Z  INFO forest::state_migration::common::state_migration: Processed 3500000 actors
2025-04-21T11:51:31.553214Z  INFO forest::state_migration::common::state_migration: Processing deferred migrations
2025-04-21T11:51:32.293837Z  INFO forest::state_migration::common::state_migration: Processed 3536266 deferred migrations
Done. old_state: bafy2bzaced5lda3j6wpzmgglta47xcxdmnxhanxyiugdme4bl3fdieiaot6go, new_state: bafy2bzacebigaxkbsthgjskbkl37fnt7c4nndqx2vbg223f6dyy6scu6izxye, took: 5m 4s 90ms 552us 64ns
➜  forest git:(main) ✗ forest-tool shed migrate-state --chain mainnet --db-write-buffer 0 25 bafy2bzacebb44z6iy77lzpkdxosxk5c2rxtyststp6ukvbknfokswi4czrr3i
Migrating... state_root: bafy2bzaced5lda3j6wpzmgglta47xcxdmnxhanxyiugdme4bl3fdieiaot6go, epoch: 4878799
2025-04-21T11:54:08.232087Z  INFO forest::state_migration::common::state_migration: Processed 100000 actors
...
2025-04-21T12:13:50.095542Z  INFO forest::state_migration::common::state_migration: Processed 3500000 actors
2025-04-21T12:13:59.860449Z  INFO forest::state_migration::common::state_migration: Processing deferred migrations
2025-04-21T12:14:00.364377Z  INFO forest::state_migration::common::state_migration: Processed 3536266 deferred migrations
Done. old_state: bafy2bzaced5lda3j6wpzmgglta47xcxdmnxhanxyiugdme4bl3fdieiaot6go, new_state: bafy2bzacebigaxkbsthgjskbkl37fnt7c4nndqx2vbg223f6dyy6scu6izxye, took: 22m 29s 894ms 378us 539ns
```

Changes introduced in this pull request:

-

## Reference issue to close (if applicable)

<!-- Include the issue reference this pull request is connected to -->
<!-- See more keywords here https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword -->
<!--(e.g. Closes #1)-->

Closes

## Other information and links

<!-- Add any other context about the pull request here. Those might be helpful links based on your investigation, relevant commits from this or other repositories or anything else -->

## Change checklist

<!-- Please add a changelog entry for your change if needed. -->
<!-- Follow this format https://keepachangelog.com/en/1.0.0/ -->

- [x] I have performed a self-review of my own code,
- [x] I have made corresponding changes to the documentation. All new code adheres to the team's [documentation standards](https://github.com/ChainSafe/forest/wiki/Documentation-practices),
- [x] I have added tests that prove my fix is effective or that my feature works (if possible),
- [x] I have made sure the [CHANGELOG][1] is up-to-date. All user-facing changes should be reflected in this document.

<!-- Thank you 🔥 -->

[1]: https://github.com/ChainSafe/forest/blob/main/CHANGELOG.md